Banking_app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["isnendyankp <isnendyankp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
poetry-dotenv-plugin = "^0.2.0"
sqlalchemy = "^2.0.28"
mysql-connector-python = "^8.3.0"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
